Rebase from 'debian/noetic/joint_state_publisher_gui'

+++ src/joint_state_publisher_gui/__init__.py
@@ -100,7 +100,7 @@ class JointStatePublisherGui(QWidget):
 
             slider.setFont(font)
             slider.setRange(0, RANGE)
-            slider.setValue(RANGE/2)
+            slider.setValue(int(RANGE/2))
 
             joint_layout.addWidget(slider)
 
@@ -159,7 +159,7 @@ class JointStatePublisherGui(QWidget):
         joint_info['slidervalue'] = joint_info['slider'].value()
         joint = joint_info['joint']
         joint['position'] = self.sliderToValue(joint_info['slidervalue'], joint)
-        joint_info['display'].setText("%.2f" % joint['position'])
+        joint_info['display'].setText("%.3f" % joint['position'])
 
     @pyqtSlot()
     def updateSliders(self):
@@ -222,7 +222,7 @@ class JointStatePublisherGui(QWidget):
         self.update_sliders()
 
     def valueToSlider(self, value, joint):
-        return (value - joint['min']) * float(RANGE) / (joint['max'] - joint['min'])
+        return int((value - joint['min']) * float(RANGE) / (joint['max'] - joint['min']))
 
     def sliderToValue(self, slider, joint):
         pctvalue = slider / float(RANGE)
